requirements.txt
couchbase==4.1.1
python-dotenv==0.21.0
tqdm==4.64.1
